move community interest tab logic into own file
